Add proper event loop for authenicate workflow.
- Removed the auth state handler fn. It did not seem like we need this since we can manage it now via the rtk query workflow.
