+++ src/App.jsx
@@ -8,13 +8,11 @@ import { TourProvider } from "@reactour/tour";
 import { buildAppRoutes } from "common/ApplicationConfig";
 import ScrollTopProvider from "common/ScrollTop/ScrollTopProvider";
 import { GeneratedTourSteps } from "common/TourSteps";
-import { fetchLoggedInUser } from "common/utils";
 import Layout from "features/Layout/Layout";
 import { MainAppRoutes } from "src/Routes";
 import { darkTheme, lightTheme } from "src/Theme";
 
 function App() {
-  const user = fetchLoggedInUser();
   const [currentThemeIdx, setCurrentThemeIdx] = useState(
     localStorage.getItem("theme") || 0,
   );
@@ -38,7 +36,7 @@ function App() {
                   />
                 }
               >
-                {buildAppRoutes(MainAppRoutes, user?.role)}
+                {buildAppRoutes(MainAppRoutes)}
               </Route>
             </Routes>
           </BrowserRouter>
