+++ src/features/Api/firebaseUserApi.js
@@ -1,10 +1,11 @@
 import secureLocalStorage from "react-secure-storage";
 
 import { createApi, fakeBaseQuery } from "@reduxjs/toolkit/query/react";
-import { authenticateViaGoogle } from "features/Auth/AuthHelper";
+import { Role, authenticateViaGoogle } from "features/Auth/AuthHelper";
 import { getAuth, signOut } from "firebase/auth";
 import {
   collection,
+  deleteDoc,
   doc,
   getDoc,
   getDocs,
@@ -85,6 +86,33 @@ export const firebaseUserApi = createApi({
             role: refetchUserData?.role,
             email: userDetails?.googleEmailAddress,
           };
+
+          // if the user has no roles, the user is a generic user
+          // generic users can have associated invites.
+          if (!Object.values(Role).includes(userWithRole.role)) {
+            const inviteRef = doc(
+              db,
+              "invites",
+              userDetails?.googleEmailAddress,
+            );
+            const inviteSnapshot = await getDoc(inviteRef);
+
+            if (inviteSnapshot.exists()) {
+              const invite = inviteSnapshot.data();
+              // Create user from invite
+              await setDoc(userRef, {
+                uid: userDetails?.uid,
+                googleEmailAddress: userDetails.googleEmailAddress,
+                googleDisplayName: userDetails.displayName ?? null,
+                googlePhotoURL: userDetails.photoURL ?? null,
+                role: invite.role,
+              });
+
+              // remove invite doc if user is created
+              await deleteDoc(inviteRef);
+              return;
+            }
+          }
 
           secureLocalStorage.setItem("user", userWithRole);
           return { data: userWithRole };
